requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
caldav
PyYAML
requests
python-dateutil
schedule
pytest
vobject
